+++ packages/twenty-front/src/modules/auth/sign-in-up/components/internal/SignInUpSSOButtonStyles.ts
@@ -1,0 +1,15 @@
+import styled from '@emotion/styled';
+import { Pill } from 'twenty-ui/components';
+
+export const StyledSSOButtonContainer = styled.div`
+  position: relative;
+  width: 100%;
+`;
+
+export const StyledLastUsedPill = styled(Pill)`
+  position: absolute;
+  right: -${({ theme }) => theme.spacing(2)};
+  top: -${({ theme }) => theme.spacing(2)};
+  background: ${({ theme }) => theme.color.blue};
+  color: ${({ theme }) => theme.font.color.inverted};
+`;
